# libs/contracts/rest/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "SetScanPolicyRequestDto": {
        "properties": {},
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "Backend/API-first social monitoring MVP.",
    "title": "Social Monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/feed/items": {
      "get": {
        "operationId": "FeedController_list",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "topicId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "description": "Optional Bearer API key. Requires read:feed. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Feed reads allow owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List tenant/workspace feed items with cursor pagination.",
        "tags": [
          "feed"
        ]
      }
    },
    "/feed/items/{feedItemId}": {
      "get": {
        "operationId": "FeedController_get",
        "parameters": [
          {
            "in": "path",
            "name": "feedItemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires read:feed. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Feed reads allow owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get one tenant/workspace feed item by id.",
        "tags": [
          "feed"
        ]
      }
    },
...
    "/source-bindings/{sourceBindingId}/scan-policy": {
      "get": {
        "operationId": "ScanPolicyController_get",
        "parameters": [
          {
            "in": "path",
            "name": "sourceBindingId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires read:topics. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Scan policy reads allow owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get scan policy for a source binding.",
        "tags": [
          "scan-policies"
        ]
      },
      "post": {
        "operationId": "ScanPolicyController_create",
        "parameters": [
          {
            "in": "path",
            "name": "sourceBindingId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires write:source_bindings. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Scan policy changes require owner or admin. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetScanPolicyRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Set scan policy for a source binding.",
        "tags": [
          "scan-policies"
        ]
      }
    },
...
  },
  "servers": [],
...
}